fix: conditional install seeed-voicecard & fix python if condition

+++ Backend/setup.sh
@@ -1,9 +1,7 @@
 sudo apt update
 sudo apt-get install -y build-essential tk-dev libncurses5-dev libnss3-dev libatlas-base-dev libncursesw5-dev libreadline6-dev libdb5.3-dev libgdbm-dev libsqlite3-dev libssl-dev libbz2-dev libexpat1-dev liblzma-dev zlib1g-dev libffi-dev
-if [type -P python3.7 > /dev/null 2>&1] 
+if [-z "$(type -P python3.7)"] 
 then
-    echo "Python3.7 is already installed"
-else
     curl -O https://www.python.org/ftp/python/3.7.3/Python-3.7.3.tar.xz
     tar -xf Python-3.7.3.tar.xz
     cd Python-3.7.3
@@ -13,6 +11,8 @@ else
     cd ../
     sudo rm -rf Python-3.7.3
     rm Python-3.7.3.tar.xz
+else
+    echo "Python3.7 is already installed"
 fi
 sudo apt-get install -y python3-gst-1.0
 sudo apt-get install -y alsa-utils
@@ -29,8 +29,13 @@ sudo ln -s /usr/bin/llvm-config-11 /usr/bin/llvm-config
 
 python3.7 -m pip install --user -r requirements.txt
 
-sudo apt-get update
-git clone https://github.com/respeaker/seeed-voicecard.git
-sudo ./seeed-voicecard/install.sh
-sudo rm -r ./seeed-voicecard
-sudo reboot
+if [ -z "$(uname -a | grep raspberrypi)"]
+then
+    sudo apt-get update
+    git clone https://github.com/respeaker/seeed-voicecard.git
+    sudo ./seeed-voicecard/install.sh
+    sudo rm -r ./seeed-voicecard
+    sudo reboot
+else
+    echo "Not a raspberrypi, stop installing seeed-voicecard package."
+fi
